client/requirements.txt
pygame==2.6.1
websocket-client==1.8.0
python-socketio==5.11.4